FS repository moved to core
