+++ knotx-example/knotx-example-app/config/application.conf
@@ -1,7 +1,7 @@
 modules = [
   "server=io.knotx.server.KnotxServerVerticle"
   "httpRepo=io.knotx.repository.HttpRepositoryConnectorVerticle"
-  "fsRepo=io.knotx.repository.FilesystemRepositoryConnectorVerticle"
+  "fsRepo=io.knotx.repository.fs.FilesystemRepositoryConnectorVerticle"
   "splitter=io.knotx.knot.splitter.FragmentSplitterVerticle"
   "assembler=io.knotx.knot.assembler.FragmentAssemblerVerticle"
   "hbsKnot=io.knotx.knot.templating.HandlebarsKnotVerticle"
