+++ knotx-example/knotx-example-app/pom.xml
@@ -49,11 +49,6 @@
     </dependency>
     <dependency>
       <groupId>io.knotx</groupId>
-      <artifactId>knotx-repository-connector-filesystem</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>io.knotx</groupId>
       <artifactId>knotx-knot-fragment-splitter</artifactId>
       <version>${project.version}</version>
     </dependency>
