--- knotx-repository-connector/knotx-repository-connector-filesystem/pom.xml
@@ -1,47 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (C) 2016 Cognifide Limited
-
-    Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-        http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-
--->
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xmlns="http://maven.apache.org/POM/4.0.0"
-    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>io.knotx</groupId>
-    <artifactId>knotx-repository-connector</artifactId>
-    <version>1.2.2-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>knotx-repository-connector-filesystem</artifactId>
-
-  <name>Knot.x - Reactive microservice assembler - Repositories Connector - Filesystem</name>
-
-  <dependencies>
-    <dependency>
-      <groupId>io.knotx</groupId>
-      <artifactId>knotx-core</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-
-    <!-- Tests -->
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-    </dependency>
-  </dependencies>
-
-</project>
