+++ tmp-core/src/main/java/io/knotx/repository/fs/FilesystemRepositoryConnectorProxyImpl.java
@@ -13,12 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.knotx.repository.impl;
+package io.knotx.repository.fs;
 
 import io.knotx.dataobjects.ClientRequest;
 import io.knotx.dataobjects.ClientResponse;
 import io.knotx.proxy.RepositoryConnectorProxy;
-import io.knotx.repository.FilesystemRepositoryOptions;
 import io.netty.handler.codec.http.HttpResponseStatus;
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Future;
